Refactored test names
