+++ integration-authorization/src/test/java/org/symphonyoss/integration/authorization/oauth/v1/OAuth1MissingParametersExceptionTest.java
@@ -3,7 +3,6 @@ package org.symphonyoss.integration.authorization.oauth.v1;
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
-import org.symphonyoss.integration.authorization.oauth.v1.exception.OAuth1MissingParametersException;
 import org.symphonyoss.integration.exception.ExpectedMessageBuilder;
 
 /**
